requirement.txt
PyMuPDF==1.24.9
pillow==10.4.0
pytesseract==0.3.13
python-doctr[torch]==0.8.1
torch>=2.0.0
torchvision>=0.15.0